First refactor : Clean the MapsFragment and put everything in a GoogleMapHelper object
