+++ app/src/main/AndroidManifest.xml
@@ -34,7 +34,7 @@
         -->
         <meta-data
             android:name="com.google.android.geo.API_KEY"
-            android:value="TODO" />
+            android:value="${MAPS_API_KEY}" />
 
         <activity android:name=".MainActivity">
             <intent-filter>
